+++ cmd/forklift/plt/deployments.go
@@ -123,3 +123,29 @@ func addDeplFeatAction(versions Versions) cli.ActionFunc {
 		return nil
 	}
 }
+
+// rm-depl-feat
+
+func rmDeplFeatAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		pallet, repoCache, _, err := processFullBaseArgs(c.String("workspace"), true)
+		if err != nil {
+			return err
+		}
+		if err = fcli.CheckShallowCompatibility(
+			pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
+			c.Bool("ignore-tool-version"),
+		); err != nil {
+			return err
+		}
+
+		deplName := c.Args().Slice()[0]
+		features := c.Args().Slice()[1:]
+		if err = fcli.RemoveDeplFeat(0, pallet, deplName, features); err != nil {
+			return err
+		}
+
+		fmt.Println("Done!")
+		return nil
+	}
+}